refactor(lib): extract shared base-url utility to avoid duplication
- Create src/lib/base-url.ts with robust URL resolution logic
- Remove duplicated getBaseUrl() from strava.ts (with unreachable branch)
- Refactor invitation-tokens.ts to use shared utility
- Centralize env var priority: APP_URL > BETTER_AUTH_URL > VERCEL URLs > dev fallback
- Add URL normalization, protocol validation, and trailing slash handling

Addresses CodeRabbit feedback about unreachable STRAVA_REDIRECT_URI check
and duplicate logic between strava.ts and invitation-tokens.ts
